We now actually send Invite Links
They don't work yet, of course, but one problem at a time.

Introduced the Config system.

+++ querki/app/identity/User.scala
@@ -61,7 +61,7 @@ object Identity {
              AND kind={kind}
           """).on("email" -> email.addr, "kind" -> IdentityKind.SimpleEmail)
       val stream = identityQuery.apply()
-      stream.headOption.map(row => Identity(OID(row.get[Int]("id").get), email)).getOrElse {
+      stream.headOption.map(row => Identity(OID(row.get[Long]("id").get), email)).getOrElse {
         val identityId = OID.next
         SQL("""
             INSERT INTO Identity
